Update README.md
This repository contains a major portion of the implementation of XAI-TMS, a novel explainable AI-based trust management system for detecting and interpreting untrustworthy behaviour in IoT environments. The system uses autoencoders, SHAP (Shapley Additive Explanations), and counterfactual generation to make trust decisions that are easy to understand, and these have been tested on standard datasets.